x/leverage: updates
